ModelClass changes
Created Model Classes for PDF content and Metadata. Also made changes to script to run off these Model Classes
